requirements.txt
accelerate==0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
auto_gptq==0.7.1+cu118
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
datasets==2.21.0
dill==0.3.8
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.112.4
ffmpy==0.4.0
filelock==3.13.1
flash-attn==2.6.3
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
gekko==1.2.1
gradio==4.43.0
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
lapx==0.5.10.post1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
modelscope==1.17.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.10.0.84
optimum==1.21.4
orjson==3.10.7
packaging==24.1
pandas==2.2.2
peft==0.12.0
pillow==10.2.0
protobuf==5.28.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyav==13.0.0
pydantic==2.9.0
pydantic_core==2.23.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
qwen-vl-utils==0.0.4
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rouge==1.0.1
ruff==0.6.4
safetensors==0.4.4
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
sympy==1.12
textblob==0.18.0.post0
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
tqdm==4.66.5
transformers @ git+https://github.com/huggingface/transformers.git@28952248b19db29ca25ccf34a5eec413376494a9
triton==3.0.0
typer==0.12.5
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.89
ultralytics-thop==2.0.6
urllib3==2.2.2
uvicorn==0.30.6
websockets==12.0
Werkzeug==3.0.4
xxhash==3.5.0
yarl==1.9.7
